+++ cmd/firearweave/cli/merger.go
@@ -17,7 +17,6 @@ func init() {
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().Duration("merger-time-between-store-lookups", 1*time.Second, "Delay between source store polling (should be higher for remote storage)")
 			cmd.Flags().Duration("merger-time-between-store-pruning", time.Minute, "Delay between source store pruning loops")
-			cmd.Flags().Uint64("merger-prune-forked-blocks-after", 50000, "Number of blocks that must pass before we delete old forks (one-block-files lingering)")
 			cmd.Flags().String("merger-grpc-listen-addr", MergerServingAddr, "Address to listen for incoming gRPC requests")
 			cmd.Flags().Uint64("merger-stop-block", 0, "if non-zero, merger will trigger shutdown when blocks have been merged up to this block")
 
@@ -30,16 +29,14 @@ func init() {
 			return nil
 		},
 		FactoryFunc: func(runtime *launcher.Runtime) (launcher.App, error) {
-			mergedBlocksStoreURL, oneBlockStoreURL, forkedBlocksStoreURL, err := getCommonStoresURLs(runtime.AbsDataDir)
+			mergedBlocksStoreURL, oneBlockStoreURL, err := getCommonStoresURLs(runtime.AbsDataDir)
 			if err != nil {
 				return nil, err
 			}
 			return mergerApp.New(&mergerApp.Config{
 				StorageOneBlockFilesPath:     oneBlockStoreURL,
 				StorageMergedBlocksFilesPath: mergedBlocksStoreURL,
-				StorageForkedBlocksFilesPath: forkedBlocksStoreURL,
 				GRPCListenAddr:               viper.GetString("merger-grpc-listen-addr"),
-				PruneForkedBlocksAfter:       viper.GetUint64("merger-prune-forked-blocks-after"),
 				StopBlock:                    viper.GetUint64("merger-stop-block"),
 				TimeBetweenPruning:           viper.GetDuration("merger-time-between-store-pruning"),
 				TimeBetweenPolling:           viper.GetDuration("merger-time-between-store-lookups"),