+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). See [MAINTAINERS.md](./MAINTAINERS.md)
 for instructions to keep up to date.
 
+## UNRELEASED
+* Removed 'forked' blocks references in the logs
+* Removed flags: --common-forked-blocks-store-url and --merger-prune-forked-blocks-after
+
 ## v1.1.1
 
 * Added 'sf.firehose.v2.Fetch' endpoint